Adding my username
When trying to clone and fork the repository I was having trouble. Basically I was unable to fetch from upstream, only the origin was shown. I have since try to delete and restart the process but it says that permission is denied in the terminal. I tried to use another repository but non were working. So basically, I was able to make changes to my clone but could not merge it here. I am going to continue to work at this and see if I am able to figure it out, but if this is something you could help me with that would be great.